Sorting Network strategies to class files now that things are working well.
