+++ BedrockService/Service/Networking/IUpdater.cs
@@ -2,7 +2,6 @@
     public interface IUpdater {
         void Initialize();
         Task CheckUpdates();
-        Task FetchBuild(string path, string version);
         bool CheckVersionChanged();
         void MarkUpToDate();
     }
